app/requirements.txt
anthropic>=0.30.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
sentence-transformers>=3.0.0
pdfplumber>=0.11.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
numpy>=1.26.0
torch>=2.2.0
transformers>=4.40.0
